+++ scripts/macros/homebrew/bg3/maimingStrike.js
@@ -2,7 +2,10 @@ import {constants} from '../../../constants.js';
 import {chris} from '../../../helperFunctions.js';
 import {queue} from '../../../utility/queue.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
+    let queueSetup = await queue.setup(workflow.item.uuid, 'maimingStrike', 50);
+    if (!queueSetup) return;
     if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+    queue.remove(workflow.item.uuid);
     if (workflow.targets.size != 1) return;
     if (!workflow.targets.first().actor.system.attributes.movement.fly) return;
     let effectData = {