+++ scripts/macros/homebrew/bg3/piercingStrike.js
@@ -3,7 +3,10 @@ import {queue} from '../../../utility/queue.js';
 import {constants} from '../../../constants.js';
 import {translate} from '../../../translations.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
+    let queueSetup = await queue.setup(workflow.item.uuid, 'piercingStrike', 50);
+    if (!queueSetup) return;
     if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+    queue.remove(workflow.item.uuid);
     if (workflow.targets.size != 1) return;
     let race = chris.raceOrType(workflow.targets.first().actor);
     if (!(race === 'undead' || race === 'construct')) return;