+++ scripts/macros/homebrew/bg3/weakeningStrike.js
@@ -1,6 +1,10 @@
 import {chris} from '../../../helperFunctions.js';
+import {queue} from '../../../utility/queue.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
+    let queueSetup = await queue.setup(workflow.item.uuid, 'topple', 50);
+    if (!queueSetup) return;
     if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+    queue.remove(workflow.item.uuid);
     if (workflow.targets.size != 1) return;
     let weapons = workflow.targets.first().actor.items.filter(i => i.system.equipped && i.type === 'weapon');
     if (weapons.length) return;